arm: tegra: soctherm: change period reg to u32.
SOC_THERM_EDP_OC_ALARM_OC4_THROTTLE_PERIOD_0 is 32bit reg.

Bug 1282441

Reviewed-on: http://git-master/r/234940

+++ arch/arm/mach-tegra/tegra11_soctherm.h
@@ -106,7 +106,7 @@ struct soctherm_throttle {
 	u8 throt_mode;
 	u8 polarity;
 	u8 priority;
-	u8 period;
+	u32 period;
 	bool intr;
 	struct soctherm_throttle_dev devs[THROTTLE_DEV_SIZE];
 };
